Move App.tsx entry file into src folder.
