+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "gymlocker",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "./src/App.tsx",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
